+++ changelog.rst
@@ -1,12 +1,15 @@
 Change Log
 ----------
 
-0.1.9 released ???
+0.1.9 released 2011-12-13
 =========================
 
 * MAJOR: fix bug that was causing groups to be deleted when a user is deleted.  Make
     sure that you run the fix-group-fk task to shore up the DB constraints.
 * make User.testing_create() more robust
+* fix bug in user permission map template that was causing exceptions if a user
+    was assigned to a group with permissions
+* add console command to add administrative user
 
 0.1.8 released 2011-11-09
 =========================